Move relevant Searchbar subcomponents to SearchResults
