+++ src/components/SearchResults/SearchResults.js
@@ -14,11 +14,11 @@ import { reportAnalytics } from '../../utils/report-analytics';
 import { getSearchbarResultsFromJSON } from '../../utils/get-searchbar-results-from-json';
 import { escapeHtml } from '../../utils/escape-reserved-html-characters';
 import { searchParamsToURL } from '../../utils/search-params-to-url';
-import SearchContext from '../Searchbar/SearchContext';
-import SearchFilters from '../Searchbar/SearchFilters';
-import SearchResult from '../Searchbar/SearchResult';
 import { useMarianManifests } from '../../hooks/use-marian-manifests';
 import Tag, { searchTagStyle } from '../Tag';
+import SearchContext from './SearchContext';
+import SearchFilters from './SearchFilters';
+import SearchResult from './SearchResult';
 import EmptyResults, { EMPTY_STATE_HEIGHT } from './EmptyResults';
 import MobileFilters from './MobileFilters';
 import 'react-loading-skeleton/dist/skeleton.css';
